requirements.txt
azure-functions
azure-functions-durable
azure-data-tables
azure-storage-blob
Pillow